- tentative mapping

+++ src/Hexapod_Serial.h
@@ -30,7 +30,7 @@ class Hexapod_Serial
 private:
     char buffer[MAX_BUF];          // where we store the message until we get a ';'
     int sofar = 0;                 // how much is in the buffer
-    unsigned int x_axis, y_axis, z_axis, roll, pitch, yaw; // positions
+    double x_axis, y_axis, z_axis, roll, pitch, yaw; // positions
     char mode_abs = 1;             // absolute mode?
     long line_number = 0;
